// src/shared/types.d.ts
interface CountryType {
  code: string;
  label: string;
  phone: string;
  suggested?: boolean;
}